Update more apps
